+++ xenon-common/src/main/java/com/vmware/xenon/common/http/netty/NettyWebSocketRequestHandler.java
@@ -299,15 +299,12 @@ public class NettyWebSocketRequestHandler extends SimpleChannelInboundHandler<Ob
                     this.host.sendRequest(Operation
                             .createDelete(service, path)
                             .setBody(body)
-                            .setReferer(
-                                    service.getUri())
+                            .setReferer(service.getUri())
                             .setCompletion(
                                     (completedOp, failure) -> {
                                         ctx.writeAndFlush(new TextWebSocketFrame(completedOp
                                                 .getStatusCode() + " " + requestId));
-                                        Utils.atomicGetOrCreate(this.serviceSubscriptions,
-                                                service.getUri(), ConcurrentSkipListSet::new)
-                                                .remove(path);
+                                        getSubscriptions(service).remove(path);
                                     }));
                     return;
                 }
@@ -355,17 +352,14 @@ public class NettyWebSocketRequestHandler extends SimpleChannelInboundHandler<Ob
                     this.host.sendRequest(Operation
                             .createPost(service, path)
                             .setBody(body)
-                            .setReferer(
-                                    service.getUri())
+                            .setReferer(service.getUri())
                             .setCompletion(
                                     (completedOp, failure) -> {
                                         ctx.writeAndFlush(new TextWebSocketFrame(completedOp
                                                 .getStatusCode() + " " + requestId));
                                         if (completedOp.getStatusCode() >= 200
                                                 && completedOp.getStatusCode() < 300) {
-                                            Utils.atomicGetOrCreate(this.serviceSubscriptions,
-                                                    service.getUri(), ConcurrentSkipListSet::new)
-                                                    .add(path);
+                                            getSubscriptions(service).add(path);
                                         }
                                     }));
                     return;
@@ -390,6 +384,11 @@ public class NettyWebSocketRequestHandler extends SimpleChannelInboundHandler<Ob
         } catch (Exception e) {
             ctx.writeAndFlush("500 " + requestId);
         }
+    }
+
+    private Set<String> getSubscriptions(WebSocketService service) {
+        return this.serviceSubscriptions.computeIfAbsent(service.getUri(),
+                k -> new ConcurrentSkipListSet<>());
     }
 
     /**