+++ xenon-common/src/test/java/com/vmware/xenon/common/TestUtils.java
@@ -18,6 +18,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
+import static org.junit.Assert.fail;
 
 import java.io.ByteArrayOutputStream;
 import java.io.UnsupportedEncodingException;
@@ -57,6 +58,8 @@ import com.esotericsoftware.kryo.Kryo;
 import com.esotericsoftware.kryo.Kryo.DefaultInstantiatorStrategy;
 import com.esotericsoftware.kryo.io.Output;
 import com.esotericsoftware.kryo.serializers.VersionFieldSerializer;
+import com.google.gson.JsonElement;
+import com.google.gson.JsonSyntaxException;
 import com.google.gson.reflect.TypeToken;
 
 import org.junit.Assert;
@@ -179,6 +182,11 @@ public class TestUtils {
         Logger.getAnonymousLogger().info("Collisions: " + collisionCount);
     }
 
+
+    private static String computeHash(byte[] content, int offset, int length) {
+        return Long.toHexString(FNVHash.compute(content, offset, length));
+    }
+
     @Test
     public void computeByteHash() throws UnsupportedEncodingException {
         CommandLineArgumentParser.parseFromProperties(this);
@@ -189,8 +197,8 @@ public class TestUtils {
         for (int i = 0; i < this.iterationCount; i++) {
             String k = "-string-" + i;
             byte[] bytes = k.getBytes(Utils.CHARSET);
-            String stringHash = Utils.computeHash(bytes, 0, bytes.length);
-            String stringHash2 = Utils.computeHash(bytes, 0, bytes.length);
+            String stringHash = computeHash(bytes, 0, bytes.length);
+            String stringHash2 = computeHash(bytes, 0, bytes.length);
             assertEquals(stringHash, stringHash2);
             assertTrue(keys.add(k));
         }
@@ -234,30 +242,6 @@ public class TestUtils {
         long e = System.nanoTime() / 1000;
         double thpt = this.iterationCount / ((e - s) / 1000000.0);
         Logger.getAnonymousLogger().info("Throughput: " + thpt);
-    }
-
-    @Test
-    public void toHexString() {
-        byte[] bytes = new byte[4];
-        bytes[0] = 0x12;
-        bytes[1] = 0x34;
-        bytes[2] = (byte) 0xAB;
-        bytes[3] = (byte) 0xCD;
-
-        String out = Utils.toHexString(bytes);
-        assertEquals("1234abcd", out);
-    }
-
-    @Test
-    public void toHexStringZeroes() {
-        byte[] bytes = new byte[4];
-        bytes[0] = 0x00;
-        bytes[1] = 0x00;
-        bytes[2] = 0x00;
-        bytes[3] = 0x00;
-
-        String out = Utils.toHexString(bytes);
-        assertEquals("00000000", out);
     }
 
     @Test
@@ -907,40 +891,40 @@ public class TestUtils {
         SystemHostInfo systemHostInfo = new SystemHostInfo();
         systemHostInfo.properties.put(SystemHostInfo.PROPERTY_NAME_OS_NAME, expected);
 
-        assertEquals(expected, Utils.getOsName(systemHostInfo));
+        assertEquals(expected, systemHostInfo.getOsName());
     }
 
     @Test
     public void testDetermineOsFamilyForWindows() throws Exception {
         final String osName = "Windows NT";
 
-        assertEquals(OsFamily.WINDOWS, Utils.determineOsFamily(osName));
+        assertEquals(OsFamily.WINDOWS, SystemHostInfo.determineOsFamily(osName));
     }
 
     @Test
     public void testDetermineOsFamilyForLinux() throws Exception {
         final String osName = "Linux";
 
-        assertEquals(OsFamily.LINUX, Utils.determineOsFamily(osName));
+        assertEquals(OsFamily.LINUX, SystemHostInfo.determineOsFamily(osName));
     }
 
     @Test
     public void testDetermineOsFamilyForMac() throws Exception {
         final String osName = "Mac OS X";
 
-        assertEquals(OsFamily.MACOS, Utils.determineOsFamily(osName));
+        assertEquals(OsFamily.MACOS, SystemHostInfo.determineOsFamily(osName));
     }
 
     @Test
     public void testDetermineOsFamilyForOther() throws Exception {
         final String osName = "TI 99/4A";
 
-        assertEquals(OsFamily.OTHER, Utils.determineOsFamily(osName));
+        assertEquals(OsFamily.OTHER, SystemHostInfo.determineOsFamily(osName));
     }
 
     @Test
     public void testDetermineOsFamilyForNull() throws Exception {
-        assertEquals(OsFamily.OTHER, Utils.determineOsFamily(null));
+        assertEquals(OsFamily.OTHER, SystemHostInfo.determineOsFamily(null));
     }
 
     @Test
@@ -988,6 +972,46 @@ public class TestUtils {
         }
 
         assertNull(Utils.getServiceUiResourcePath(new MyService()));
+    }
+
+    @Test
+    public void testGetFromPrimitives() {
+        Map<String, Object> map = new HashMap<>();
+        map.put("bt", true);
+        map.put("bf", false);
+        map.put("int", 123);
+        map.put("double", 3.14);
+        map.put("string", "hello");
+        ExampleServiceState doc = new ExampleServiceState();
+        doc.documentSelfLink = "selfLink";
+        doc.tags = new HashSet<>(Arrays.asList("t1", "t2"));
+        map.put("doc", doc);
+
+        validateExtract(Utils.toJson(map));
+        validateExtract(Utils.fromJson(Utils.toJson(map), JsonElement.class));
+    }
+
+    private void validateExtract(Object jsonRepr) {
+        assertEquals(true, Utils.getJsonMapValue(jsonRepr, "bt", Boolean.class));
+        assertEquals(false, Utils.getJsonMapValue(jsonRepr, "bf", Boolean.class));
+        assertEquals(Integer.valueOf(123), Utils.getJsonMapValue(jsonRepr, "int", Integer.class));
+        assertEquals(Double.valueOf(3.14), Utils.getJsonMapValue(jsonRepr, "double", Double.class));
+        assertEquals("hello", Utils.getJsonMapValue(jsonRepr, "string", String.class));
+        assertEquals("selfLink", Utils.getJsonMapValue(jsonRepr, "doc", ExampleServiceState.class).documentSelfLink);
+        assertTrue(Utils.getJsonMapValue(jsonRepr, "doc", ExampleServiceState.class).tags.contains("t1"));
+
+        assertNull(Utils.getJsonMapValue(jsonRepr, "badKey", ExampleServiceState.class));
+
+        // coercion to string always possible
+        assertEquals("true", Utils.getJsonMapValue(jsonRepr, "bt", String.class));
+
+        try {
+            // must fail if object is expected
+            Utils.getJsonMapValue(jsonRepr, "bt", ServiceDocument.class);
+            fail("Impossible conversion");
+        } catch (JsonSyntaxException ignore) {
+
+        }
     }
 
     @Test