+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## 1.5.0-SNAPSHOT
 
+* Remove `Utils.computeHash`, `Utils.getCurrentFileDirectory`, `Utils.atomicGetOrCreate`,
+  `Utils.toHexString`, `Utils.toServiceErrorResponseJson` as there are
+  obvious/shorter/standard ways to achieve the same. Methods `Utils.getOsName` and
+  `Utils.determineOsFamily` are moved to `SystemHostInfo`.
+
 ## 1.4.2
 
 * Add ability to expand only selected fields in queries